PROJETOS/20241013-migracao-wallabag/requirements.txt
babel==2.16.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-repl==0.3.0
click-spinner==0.1.10
colorama==0.4.6
colorlog==6.8.2
Delorean==1.0.0
humanize==4.11.0
idna==3.10
lxml==5.3.0
markdownify==0.13.1
packaging==24.1
pocket==0.3.6
pockexport @ git+https://github.com/karlicoss/pockexport@35651edecbc45d06b1ef29fb044d431f0acabc4b
prompt_toolkit==3.0.48
pycryptodome==3.21.0
python-dateutil==2.9.0.post0
pytz==2024.2
pyxdg==0.28
requests==2.32.3
six==1.16.0
soupsieve==2.6
tabulate==0.9.0
tqdm==4.66.5
tzlocal==5.2
urllib3==2.2.3
wallabag-client==1.8.5
wcwidth==0.2.13
